requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
python-dateutil==2.8.2 